Fix various compiler warnings. (#238)
Fixed :
- Arithmetic between different enum types by casting (GCC/CLang warnings).
- Fix unneeded semicolumn after endif (GCC warning).
- Fix static method called from object. (CppCheck)
-

+++ playerbot/strategy/ReactionEngine.cpp
@@ -243,7 +243,7 @@ bool ReactionEngine::ListenAndExecute(Action* action, Event& event)
         out << " [reaction]";
 
         if(actionExecuted)
-            out << " (duration: " << ((float)incomingReaction.GetDuration() / IN_MILLISECONDS) << "s)";
+            out << " (duration: " << ((float)incomingReaction.GetDuration() / static_cast<float>(IN_MILLISECONDS)) << "s)";
 
         ai->TellPlayerNoFacing(ai->GetMaster(), out);
     }
